config file argument introduced

+++ main_test.c
@@ -34,7 +34,7 @@ void test_entropy() {
     unsigned char* s = "Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.";
     //char* s = "abcdefghilmnopqrstuvzxy1234567890";
 
-    double H= calc_rand_idx(s, strlen(s));
+    double H= calc_rand_idx(s, (long) strlen(s));
 
     if (H-4.022379<0.00001)
         printf("H: %f - SUCCESS", H);
